refactor: useReaction 훅 분리 및 sessionId 없으면 early return하도록 수정
- 테스트 코드에서 sessionId 없을 때 동작 테스트 제거
