+++ frontend/src/hooks/session/useMediaDevices.ts
